plugins: migrate six- to nine-surface modules to the `props` mapper
Consolidates plugin-registry's four category surfaces onto one
`RegistryCategoryArticle` taking the category as a prop, and moves its two
in-module helpers into that file.

plugin-debug's hook-bearing surfaces move to `DebugSurfaces.tsx`; plugin-script's
to `ScriptSurfaces.tsx`; plugin-preview's two composed cards (card plus a nested
`Related` surface) to `RelatedCards.tsx`.

plugin-preview's two explicit `Surface.create<T>` type arguments are dropped: an
explicit argument pins the data type and forces the mapper to return the whole
envelope, whereas inferring it from `filter` is the point of the mapper.

plugin-connector's selector field is typed against `Surface.ComponentProps`,
replacing the last `as any as FormFieldRendererProps`.
